+++ CHANGELOG.md
@@ -17,9 +17,10 @@
 - Added warnings to notify users if they submit an incomplete diagram when doing the fill in the diagram exercise
 - Added responsive components to diagram page
 - Changed the colour of `ComponentPieces` to be based on its placement in the diagram by the user, with sidebar pieces being a neutral colour
+- Added a new `MatchingExercise` page and its styling
 
 ### Bug Fix
 
-
 ### Internal Change
 
+- Refactored `ComponentPieces` and `ExerciseBoardSubheader` into a common folder